+++ Render_Engine_Only/ch9_planes/rt_planes.c
@@ -17,7 +17,7 @@ void	rt_ray_plane_x(t_object plane, t_ray ray, t_xs *xs, int *i)
 	float		t;
 
 	errno = 0;
-	if (fabs(ray.direction.y) < EPSILON)
+	if (fabsf(ray.direction.y) < EPSILON)
 		errno = ERAYPARALLEL;
 	else
 	{
